fix: Pretendard 적용 문제 해결 #21 (#22)
