+++ src/styles/font.ts
@@ -1,0 +1,8 @@
+import localFont from 'next/font/local';
+
+export const pretendard = localFont({
+  src: '../assets/fonts/PretendardVariable.woff2',
+  display: 'swap',
+  weight: '100 900',
+  variable: '--font-pretendard',
+});
